mcs/class/System.Web.DynamicData/System.Web.DynamicData.ModelProviders/TableProvider.cs
//
// TableProvider.cs
//
// Author:
//	Atsushi Enomoto <atsushi@ximian.com>
//      Marek Habersack <mhabersack@novell.com>
//
// Copyright (C) 2008-2009 Novell Inc. http://novell.com
//

//
// Permission is hereby granted, free of charge, to any person obtaining
// a copy of this software and associated documentation files (the
// "Software"), to deal in the Software without restriction, including
// without limitation the rights to use, copy, modify, merge, publish,
// distribute, sublicense, and/or sell copies of the Software, and to
// permit persons to whom the Software is furnished to do so, subject to
// the following conditions:
// 
// The above copyright notice and this permission notice shall be
// included in all copies or substantial portions of the Software.
// 
// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
//
using System;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel;
using System.Linq;
using System.Security.Permissions;
using System.Web.UI;

namespace System.Web.DynamicData.ModelProviders
{
	[AspNetHostingPermission (SecurityAction.LinkDemand, Level = AspNetHostingPermissionLevel.Minimal)]
	[AspNetHostingPermission (SecurityAction.InheritanceDemand, Level = AspNetHostingPermissionLevel.Minimal)]
	public abstract class TableProvider
	{
		protected TableProvider (DataModelProvider model)
		{
			DataModel = model;
		}

		public abstract ReadOnlyCollection<ColumnProvider> Columns { get; }

		public DataModelProvider DataModel { get; private set; }
		public virtual Type EntityType { get; protected set; }
		public virtual string Name { get; protected set; }

		public virtual object EvaluateForeignKey (object row, string foreignKeyName)
		{
			return DataBinder.GetPropertyValue (row, foreignKeyName);
		}

		public abstract IQueryable GetQuery (object context);

		public override string ToString ()
		{
			string name = Name;
			if (String.IsNullOrEmpty (name))
				return base.ToString ();

			return name;
		}
	}
}
